pyproject.toml
[project]
name = "dentist-appointment-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-core>=1.2.26",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-xai>=1.2.2",
    "langgraph>=1.1.6",
    "ollama>=0.6.1",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.43.0",
]
